+++ __test__/expected/issue_11.js
@@ -1,0 +1,6 @@
+import monaco from 'non-exists';
+export default {
+  foo() {
+    return monaco.Promise.wrap(import('./tsMode.js'));
+  },
+}