+++ __test__/fixtures/issue_11.js
@@ -1,0 +1,7 @@
+import monaco from 'non-exists';
+
+export default {
+  foo() {
+    return monaco.Promise.wrap(import('./tsMode.js'));
+  },
+}