+++ jest.config.js
@@ -2,5 +2,5 @@ module.exports = {
   collectCoverage: true,
   coverageDirectory: 'coverage',
   coverageReporters: ['lcov', 'text'],
-  watchman: false
+  watchman: false,
 }