+++ rollup.config.js
@@ -1,6 +1,7 @@
-
-import buble from 'rollup-plugin-buble'
-
+/*
+  Rollup config.
+  With node v6.x the bublé plugin is not necessary.
+*/
 const external = Object
   .keys(require('./package.json').dependencies)
   .concat('fs', 'path')
@@ -10,15 +11,12 @@ const banner = [
   ' * rollup-plugin-cleanup v' + require('./package.json').version,
   ' * @author aMarCruz',
   ' * @license MIT',
-  ' */'
+  ' */',
 ].join('\n')
 
 export default {
   input: 'src/index.js',
   plugins: [
-    buble({
-      target: { node: 4 }
-    })
   ],
   external,
   output: [
@@ -33,6 +31,6 @@ export default {
       format: 'es',
       banner,
       interop: false,
-    }
-  ]
+    },
+  ],
 }