+++ src/remove-lines.js
@@ -1,14 +1,18 @@
 
 import acorn from 'acorn'
 
-const EOL_TYPES   = { unix: '\n', mac: '\r', win: '\r\n' }
-const FIRST_LINES = /^\s*[\r\n]/
-const EACH_LINE   = /.*(?:\r\n?|\n)/g
-const TRIM_SPACES = /[^\S\r\n]+$/
+const EOL_TYPES = { unix: '\n', mac: '\r', win: '\r\n' }
+
+// Matches empty lines at the start of the buffer
+const FIRST_EMPTY_LINES = /^\s*[\r\n]/
+// Searches lines and captures its line-endings (of any type).
+const EACH_LINE = /.*(?:\r\n?|\n)/g
+// Matches spaces after the last line-ending
+const TRAILING_SPACES = /[^\S\r\n]+$/
 
 export default function removeLines(magicStr, code, file, options) {
 
-  // matches one or more line endings and their leading spaces
+  // Local regex that matches one or more line endings and its leading spaces
   const NEXT_LINES = /\s*[\r\n]/g
 
   const eolTo   = EOL_TYPES[options.normalizeEols]
@@ -17,13 +21,18 @@ export default function removeLines(magicStr, code, file, options) {
   // middle lines count one more
   const maxEolChars = maxEolCharsAtStart + eolTo.length
 
-  let match, block, region
   let changes = false
   let pos = 0
 
   // Helpers
   // -------
 
+  /**
+   * Replaces a string block with a new one through MagicString.
+   * @param {string} str Original block
+   * @param {number} start Offset of the block
+   * @param {string} rep New block
+   */
   const replaceBlock = (str, start, rep) => {
     if (str !== rep) {
       magicStr.overwrite(start, start + str.length, rep)
@@ -31,6 +40,13 @@ export default function removeLines(magicStr, code, file, options) {
     }
   }
 
+  /**
+   * Normalizes and compacts a block of blank characters to convert it into a
+   * block of line-endings that do not exceed the maximum number defined by the
+   * user.
+   * @param {string} str Block of blank characters to search on
+   * @param {number} max Maximum number of *characters* for the empty lines
+   */
   const limitLines = (str, max) => {
     let ss = str.replace(EACH_LINE, eolTo)
     if (ss.length > max) {
@@ -39,17 +55,32 @@ export default function removeLines(magicStr, code, file, options) {
     return ss
   }
 
+  /**
+   * Normalizes and compacts the region of the buffer bounded by `start` and `end`.
+   * @param {number} start Offset of the start of the region
+   * @param {number} end Ending of the region
+   * @param {boolean} atStart We are at the beginning of the buffer?
+   * @param {boolean} atEnd At the end of the buffer?
+   */
   const squashRegion = (start, end, atStart, atEnd) => {
-    NEXT_LINES.lastIndex = 0
-    region = magicStr.slice(start, end)
+    const region = magicStr.slice(start, end)
+    let block
+    let match
 
     // first empty lines
-    if (atStart && (match = region.match(FIRST_LINES))) {
+    if (atStart && (match = region.match(FIRST_EMPTY_LINES))) {
       block = match[0]
       replaceBlock(block, start, limitLines(block, maxEolCharsAtStart))
+
+      // Set lastIndex to the start of the first non-empty line in this region
       NEXT_LINES.lastIndex = block.length
+    } else {
+      // Reset lastIndex to the start of this region
+      NEXT_LINES.lastIndex = 0
     }
 
+    // Compact intermediate lines, if `maxEmptyLines` is zero all blank lines
+    // are removed. If it is -1 the spaces are removed, keeping the EOLs.
     if (empties) {
       // maxEmptyLines -1 or > 0
       while ((match = NEXT_LINES.exec(region))) {
@@ -63,7 +94,8 @@ export default function removeLines(magicStr, code, file, options) {
       }
     }
 
-    if (atEnd && (match = TRIM_SPACES.exec(region))) {
+    // Cut the spaces after the final EOL
+    if (atEnd && (match = TRAILING_SPACES.exec(region))) {
       replaceBlock(match[0], start + match.index, '')
     }
   }
@@ -79,8 +111,7 @@ export default function removeLines(magicStr, code, file, options) {
   // --------------
 
   acorn.parse(code, {
-    ecmaVersion: options.ecmaVersion,
-    sourceType: options.sourceType,
+    ...options.acornOptions,
     onToken,
   })