+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # rollup-plugin-cleanup changes
 
 ### v3.0.0
+- Add an example of using Acorn plugin the README.
+- New option `acornOptions`, for advanced usage.
+  This is an optional, plain JS object with additional settings passed to the [Acorn](https://github.com/acornjs/acorn) parser. Properties of this object are merged with, and take precedence over, the existing `ecmaVersion` and `sourceType` options.
+- The mimimum node version is now 6.14, for compatibility with ESLint 5.
+- Updated devDependencies.
+
+### v3.0.0-beta.1
 - Removed the [riot](http://riotjs.com/) .tag extension from the defaults, you can add this manually.
 - Now acorn `ecmaVersion` defaults to 9 (ES2018).
 - Closes #10 : Errors out on spread operator.